+++ constance/management/commands/constance.py
@@ -1,4 +1,3 @@
-from django.apps import apps
 from django.conf import settings
 from django.core.exceptions import ValidationError
 from django.core.management import BaseCommand, CommandError
@@ -8,6 +7,7 @@ from django import VERSION
 
 from ... import config
 from ...admin import ConstanceForm, get_values
+from ...models import Constance
 
 
 def _set_constance_value(key, value):
@@ -49,7 +49,6 @@ class Command(BaseCommand):
             help='delete all Constance keys and their values if they are not in settings.CONSTANCE_CONFIG (stale keys)',
         )
 
-
     def _subparsers_add_parser(self, subparsers, name, **kwargs):
         # API in Django >= 2.1 changed and removed cmd parameter from add_parser
         if VERSION >= (2, 1) and 'cmd' in kwargs:
@@ -82,23 +81,16 @@ class Command(BaseCommand):
                 self.stdout.write("{}\t{}".format(k, v), ending="\n")
 
         elif command == 'remove_stale_keys':
-            try:
-                Constance = apps.get_model('database.Constance')
-            except LookupError:
-                Constance = None
 
-            if Constance:
-                actual_keys = settings.CONSTANCE_CONFIG.keys()
+            actual_keys = settings.CONSTANCE_CONFIG.keys()
 
-                stale_records = Constance.objects.exclude(key__in=actual_keys)
-                if stale_records:
-                    self.stdout.write("The following record will be deleted:", ending="\n")
-                else:
-                    self.stdout.write("There are no stale records in database.", ending="\n")
-
-                for stale_record in stale_records:
-                    self.stdout.write("{}\t{}".format(stale_record.key, stale_record.value), ending="\n")
-
-                stale_records.delete()
+            stale_records = Constance.objects.exclude(key__in=actual_keys)
+            if stale_records:
+                self.stdout.write("The following record will be deleted:", ending="\n")
             else:
-                self.stdout.write("Database backend is not set. Nothing is deleted", ending="\n")
+                self.stdout.write("There are no stale records in database.", ending="\n")
+
+            for stale_record in stale_records:
+                self.stdout.write("{}\t{}".format(stale_record.key, stale_record.value), ending="\n")
+
+            stale_records.delete()