+++ example/cheeseshop/settings.py
@@ -44,7 +44,6 @@ INSTALLED_APPS = (
     'cheeseshop.apps.catalog',
     'cheeseshop.apps.storage',
     'constance',
-    'constance.backends.database',
 )
 
 MIDDLEWARE = (
@@ -149,3 +148,5 @@ USE_TZ = True
 # https://docs.djangoproject.com/en/1.8/howto/static-files/
 
 STATIC_URL = '/static/'
+
+DEFAULT_AUTO_FIELD = 'django.db.models.AutoField'