+++ tests/test_checks.py
@@ -1,6 +1,6 @@
 import datetime
 from decimal import Decimal
-import mock
+from unittest import mock
 
 from constance.admin import get_values
 from constance.checks import check_fieldsets, get_inconsistent_fieldnames